arousal/arousal_requirements.txt
absl-py==1.2.0
aiosignal==1.3.1
antropy==0.1.4
appdirs==1.4.4
asttokens==2.2.0
astunparse==1.6.3
attrs==23.1.0
audioread==2.1.9
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.0
click==8.1.3
comm==0.1.1
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
distlib==0.3.6
dm-tree==0.1.8
entrypoints==0.4
executing==1.2.0
filelock==3.12.0
flatbuffers==2.0
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2023.10.0
gast==0.5.3
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
hdf5storage==0.1.18
heartpy==1.2.7
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.18.2
ipython==8.7.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-widgets==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.4
libclang==14.0.1
librosa==0.9.2
littleutils==0.2.2
llvmlite==0.39.0
lspopt==1.2.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mat4py==0.5.0
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mne==1.7.1
mpl-point-clicker==0.3.1
mpmath==1.3.0
msgpack==1.0.5
namex==0.0.7
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8
numba
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
operator-courier==2.1.11
operators==1.0.1
opt-einsum==3.3.0
outdated==0.2.2
packaging==24.1
panda==0.3.1
pandas==2.2.2
pandas-flavor==0.6.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pingouin==0.5.4
platformdirs==2.5.4
pooch==1.6.0
prompt-toolkit==3.0.33
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyEDFlib==0.1.30
Pygments==2.13.0
pykeyboard==0.1.5
pyparsing==3.0.9
pyriemann==0.3
Pyrogram==2.0.57
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-xlib==0.31
pytz==2022.2
PyUserInput==0.1.11
PyYAML==6.0
pyzmq==24.0.1
ray==2.4.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy
rich==13.7.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy
seaborn==0.12.1
semver==3.0.2
six==1.16.0
sleepecg==0.5.2
SoundFile==0.10.3.post1
stack-data==0.6.2
statsmodels==0.14.2
stochastic
sympy==1.12
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
tensorpac==0.6.5
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
TgCrypto==1.2.4
threadpoolctl==3.1.0
tk
tornado==6.2
tqdm==4.66.4
traitlets==5.6.0
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.3.0
tzdata==2024.1
urllib3==1.26.10
validators==0.22.0
virtualenv==20.21.0
wcwidth==0.2.5
Werkzeug==2.1.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xarray==2024.6.0
xlib==0.21
yasa==0.6.2
zipp==3.8.1
